+++ recognition/_evaluation_/ijb/ijb_11.py
@@ -51,8 +51,8 @@ job = args.job
 def read_template_media_list(path):
     #ijb_meta = np.loadtxt(path, dtype=str)
     ijb_meta = pd.read_csv(path, sep=' ', header=None).values
-    templates = ijb_meta[:, 1].astype(np.int)
-    medias = ijb_meta[:, 2].astype(np.int)
+    templates = ijb_meta[:, 1].astype(np.int32)
+    medias = ijb_meta[:, 2].astype(np.int32)
     return templates, medias
 
 
@@ -63,10 +63,10 @@ def read_template_pair_list(path):
     #pairs = np.loadtxt(path, dtype=str)
     pairs = pd.read_csv(path, sep=' ', header=None).values
     #print(pairs.shape)
-    #print(pairs[:, 0].astype(np.int))
-    t1 = pairs[:, 0].astype(np.int)
-    t2 = pairs[:, 1].astype(np.int)
-    label = pairs[:, 2].astype(np.int)
+    #print(pairs[:, 0].astype(np.int32))
+    t1 = pairs[:, 0].astype(np.int32)
+    t2 = pairs[:, 1].astype(np.int32)
+    label = pairs[:, 2].astype(np.int32)
     return t1, t2, label
